Rename TranscriptSearch to KeywordSearch; move Hit type to SpokenSearch
