+++ Sources/PalmierPro/MediaPanel/MediaTab/MediaTab+Search.swift
@@ -115,7 +115,7 @@ extension MediaTab {
         }
     }
 
-    private func spokenRow(_ hit: TranscriptSearch.Hit) -> some View {
+    private func spokenRow(_ hit: SpokenSearch.Hit) -> some View {
         let asset = editor.mediaAssets.first { $0.id == hit.assetID }
         let range = hit.start...max(hit.end, hit.start + 0.1)
         let thumbW = CGFloat(thumbnailSize) * 1.4
